Nunito MICRO update
revised and updated all of the numerals: I made all numerals lighter in Heavy opsz
I made superiors taller but not wider. In ExtraLight 10 units taller (so just the amount we raised the x-height) and in Heavy 18 units (same principle, x-height). I think this is a good logical start. I think this is a good start and they are already better legible in super small sizes than before.
I have still different top of 6 and 6.tosf in ExtraLight — I will clean this… I think that curve in the six.tosf is nicer than the one on the six